requirements.txt
aiohttp>=3.9
aioredis>=2
asyncpg
discord
grpcio>=1.53.2,<=1.53.2
hypercorn
icalendar
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-aiohttp-client
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-asyncpg
opentelemetry-instrumentation-dbapi
opentelemetry-instrumentation-jinja2
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-system-metrics
python-dateutil
python-jose
quart>=0.19
quart-session
redis
SQLAlchemy>=2
urllib3>=1.26,<2.0
uvicorn
